requirements.txt
django~=3.2.5
djangorestframework~=3.12.4
psycopg2-binary

django-filter~=2.4.0